scripts/requirements.txt
numpy>=1.26
pillow>=10.2
matplotlib>=3.8
pycocotools>=2.0.7
